+++ eSlayerServer/src/eblockaction.cpp
@@ -10,11 +10,11 @@ eBlockAction::sCreate(
     const auto& data = unit.data();
     const int anim = data.animId("block");
     if(anim != -1) {
-        const double fbr = unit.fasterBlockRate();
-        const double efbr = std::floor(fbr*120./(fbr + 120.));
+        const float fbr = unit.fasterBlockRate();
+        const float efbr = std::floor(fbr*120.f/(fbr + 120.f));
         const int blockBase = data.animFrames(anim);
-        const double animSpeed = 256.;
-        const int frames = int(std::ceil(256.*blockBase/std::floor(animSpeed*(50. + efbr)/100.))) - 1;
+        const float animSpeed = 256.f;
+        const int frames = int(std::ceil(256.f*blockBase/std::floor(animSpeed*(50.f + efbr)/100.f))) - 1;
         const auto a = std::make_shared<eBlockAction>(unit, area);
         a->setup(anim, frames, nullptr);
         return a;
