+++ eSlayerServer/src/emovetoenemyaction.h
@@ -9,11 +9,11 @@ class eMoveToEnemyAction : public eUnitAction {
 public:
     using eUnitAction::eUnitAction;
 
-    void increment(const double by) override;
+    void increment(const float by) override;
 private:
     void setTarget(const eServerUnit& u);
     int mTargetId = -1;
-    ePointF mTargetPos{0., 0.};
+    ePointF mTargetPos{0.f, 0.f};
 };
 
 #endif // EMOVETOENEMYACTION_H
